github_cloner/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
gitpython==3.1.42
httpx==0.27.0
pydantic==2.6.3
